[usb] Hide usb mode selector dialog on mode transition. JB#42756
The rules for hiding usb mode selector are incorrect and can lead to
hiding the dialog prematurely / leaving it on screen despite it being
impossible to choose usb mode anymore.

Ignore events that are not related to mode selector dialog to avoid
premature hiding, and hide the dialog when transition to some actual
usb mode is seen.

Requires functionality that is available in libusb-moded-qt >= 1.7

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>