[lipstick] Allow control of VPN credential storage/retrieval. Contributes to JB#45178
When using the VPN the UI can request credentials from the user. These
can optionally be stored if the user selects the option to store them in
the dialogue.

There may be times when the VPN should prevent the user from choosing to
store credentials. This change allows connman to request for them not to
be stored.

The controlling input comes from connman. The resulting UI is handled by
lipstick-jolla-home.