[batterynotifier] Use -1 for unknown battery level. Fixes JB#48562
A false positive "not enough power" notification is triggered if
lipstick is startup happens while charger is connected and battery
level is below 49 percent.

Use -1 percent as initial battery level guess instead of 50 percent
(this is also what mce reports when it does not know battery level).

Adapt code that deals with battery levels so that notifications are
not triggered while battery level is unknown.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>