[lipstick] Don't keep restored state on updated notifications. JB#61050
If an application was fetching notifications via GetNotifications(),
only updating some properties, and then triggering an update,
it was possible that the internal x-nemo-restored hint leaked to the
application side and came back on the update, skipping feedbacks and
notification popup on the update. And then finally being even saved
on the notification database.

So let's once and for all make this restored status an internal
detail of the lipstick notification class.