[lipstick] Support storing privileged source of notifications. JB#60924
Introducing a separate hint set for internal properties. Ensuring
that they never get even by accident over D-Bus.

Used that to store whether the source of the notification is a process
with privileged group and to indicate that a notification action can be
trusted.

Notification database version schema not bumped as this didn't change
any existing table.